requirements.txt
aiofiles==24.1.0
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
bidict==0.23.1
blinker==1.8.2
cachelib==0.13.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cs50==9.3.4
docutils==0.19
exceptiongroup==1.2.2
fastapi==0.109.2
Flask==3.0.3
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
ifaddr==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
markdown2==2.5.0
MarkupSafe==2.1.5
msgspec==0.18.6
multidict==6.0.5
nicegui==1.4.29
orjson==3.10.6
packaging==24.0
pillow==10.3.0
pillow_heif==0.16.0
pscript==0.7.7
pydantic==1.10.17
Pygments==2.18.0
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.3
PyYAML==6.0.1
requests==2.32.3
simple-websocket==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.30
sqlparse==0.5.0
starlette==0.36.3
termcolor==2.4.0
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
vbuild==0.8.2
watchfiles==0.22.0
websockets==12.0
Werkzeug==3.0.3
wsproto==1.2.0
yarl==1.9.4
